Remove AVX check implementation
Deleted AVX check source and header files, and associated compilation steps in the Dockerfile. This simplifies the overall codebase by eliminating deprecated or unnecessary AVX detection functionality. Adjusted main.go to remove related C import references.
