--- avx/avx_check.h
@@ -1,8 +1,0 @@
-#ifndef AVX_CHECK_H
-#define AVX_CHECK_H
-
-int check_avx_support();
-int check_avx512_support();
-int check_neon_support();
-
-#endif // AVX_CHECK_H
